requirements.txt
reportlab
pymupdf
fastapi==0.109.2
uvicorn[standard]==0.27.1

pydantic-settings>=2.10.1,<3.0.0
python-dotenv==1.0.1
email-validator==2.1.0
httpx==0.27.0
python-multipart

openai==1.12.0
google-genai

pillow==10.4.0

sqlalchemy>=2.0
psycopg2-binary
asyncpg
pgvector

bcrypt
python-jose[cryptography]
tzdata

celery==5.3.6
redis==5.0.4
openpyxl
python-docx
google-cloud-storage
google-generativeai
